requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
langchain
langchain-openai
langchain-community
faiss-cpu
